Add an ant/Flex build script.  Changed the key while testing, and as it turns, out, adobe won't let you install on top of the other one, if it already exists, IF it was generated with a different key.  Manually remove the previous air app, and the next will install just fine, after warning you about my skeevy 'self signed' certificate.  The script will make a new one, if you want to build it for yourself.
Add tools to menu to clean up the mess we made (and 'are you sure' prompts, to be sure we meant to) with thumbnails.  Won't delete thumbnails that don't have an 'MP4' file associated with them.  Because it was kind of dangerous to jpegs in the wild, if you weren't careful with the path.

When we 'give up' on thumbnail quality, settle for the best one we had (as far as the dumb heuristic cares).

Settle on 240 pixels as the 'right' default thumbnail resolution (evenly divisible by 720, not 'tiny' like 180, not huge, like 360)