Video player bugs
Mostly correcting for possibly transient behaviors in Firefox.  On
full-screen mode, the player window wouldn’t center, like all of the
other browsers do.  Adjusting it would break all of the other browsers.
 So made an exceptional case for Firefox.  Also, hide play list while
in full screen, as Firefox leaves that visible, behind the playback
window and controls.

Fix a bug that crept in, that could hide the player controls, and leave
them hidden, if you hit ‘escape’ while the mouse+controls were
invisible.