More tinkering with Firefox centering
The CSS thing appeared to work, but it wasn’t working.

So back to the javascript, detect if the window isn’t centered, and
force it to be centered if it’s stuck to an edge, and not big enough to
fill that dimension.

Add a dim border around the red seek bar.

Call ‘MouseMoved’ when playback starts a new track, to make sure the
title is briefly visible.