Make a new video player front end
Added ‘kiosk’ mode skin.  Shows either list of tracks, or video, but
not both.  Makes navigating without going back and forth through ‘full
screen’ mode changes easier.

Refactored code in this skin, to dump everything into a ‘vid’
namespace, so debugging it is a bit easier.

I probably need to add a ‘back’ icon, since that’s not available when
in a ‘presentation’ or ‘kiosk’ mode.

I’ll tinker with it a bit before deciding whether to make it the
‘default’ player.