Add hierarchical indent to ‘Folder Index’ in python scripts
While not identical to the AIR app’s behavior of constructing
sub-folder links, it is very usable in this state.  I wonder whether I
should recursively call the index construction on all found folders,
and be done with it?

Through the miracle of testing, discovered I was emitting backslashes
in web URIs in the python scripts.  Fixed that.

Add ‘<meta charset="UTF-8”>’, where it was missing in index templates.