Defer sort, to get indexing started a little faster
If ‘sorted(os.walk…)’, the entire folder tree must be crawled before
anything happens.  Delaying that sort allows the individual play list
folders to be built, while slow NAS access happens in the background.
Instead of a long wait, and then some output, get output almost
immediately.  A small speed-up.

Also, officially mothballing the ‘shell only’ tools.  If you have the
shell tools, you almost certainly have Python, and if you don’t, Python
is readily available.  The dependency for perl in the shell script is
as much a dependency as Python is.