Fix some vexing full-screen issues in Chrome
When requestFullscreen was called on video box or window element, it
would turn half the top half of the window black.  When done to the
document.body, the left half horizontally, when done to the
document.documentElement, with an ugly border.  Then tracked down and
got rid of the border.

Does not fix the iOS ‘sand in the scrolling mechanism’ bug, that
happens when audio or video player is present.  That’s in Safari.  Nor
does it fix the ‘I don’t know where to be’ bug on the audio player
controls, in Silk.  That’s in Silk.